+++ components/doc/cascadeselect/filleddoc.js
@@ -1,0 +1,316 @@
+import { DocSectionCode } from '@/components/doc/common/docsectioncode';
+import { DocSectionText } from '@/components/doc/common/docsectiontext';
+import { CascadeSelect } from '@/components/lib/cascadeselect/CascadeSelect';
+import { useState } from 'react';
+
+export function FilledDoc(props) {
+    const [selectedCity, setSelectedCity] = useState(null);
+    const countries = [
+        {
+            name: 'Australia',
+            code: 'AU',
+            states: [
+                {
+                    name: 'New South Wales',
+                    cities: [
+                        { cname: 'Sydney', code: 'A-SY' },
+                        { cname: 'Newcastle', code: 'A-NE' },
+                        { cname: 'Wollongong', code: 'A-WO' }
+                    ]
+                },
+                {
+                    name: 'Queensland',
+                    cities: [
+                        { cname: 'Brisbane', code: 'A-BR' },
+                        { cname: 'Townsville', code: 'A-TO' }
+                    ]
+                }
+            ]
+        },
+        {
+            name: 'Canada',
+            code: 'CA',
+            states: [
+                {
+                    name: 'Quebec',
+                    cities: [
+                        { cname: 'Montreal', code: 'C-MO' },
+                        { cname: 'Quebec City', code: 'C-QU' }
+                    ]
+                },
+                {
+                    name: 'Ontario',
+                    cities: [
+                        { cname: 'Ottawa', code: 'C-OT' },
+                        { cname: 'Toronto', code: 'C-TO' }
+                    ]
+                }
+            ]
+        },
+        {
+            name: 'United States',
+            code: 'US',
+            states: [
+                {
+                    name: 'California',
+                    cities: [
+                        { cname: 'Los Angeles', code: 'US-LA' },
+                        { cname: 'San Diego', code: 'US-SD' },
+                        { cname: 'San Francisco', code: 'US-SF' }
+                    ]
+                },
+                {
+                    name: 'Florida',
+                    cities: [
+                        { cname: 'Jacksonville', code: 'US-JA' },
+                        { cname: 'Miami', code: 'US-MI' },
+                        { cname: 'Tampa', code: 'US-TA' },
+                        { cname: 'Orlando', code: 'US-OR' }
+                    ]
+                },
+                {
+                    name: 'Texas',
+                    cities: [
+                        { cname: 'Austin', code: 'US-AU' },
+                        { cname: 'Dallas', code: 'US-DA' },
+                        { cname: 'Houston', code: 'US-HO' }
+                    ]
+                }
+            ]
+        }
+    ];
+
+    const code = {
+        basic: `
+<CascadeSelect variant="filled" value={selectedCity} onChange={(e) => setSelectedCity(e.value)} options={countries} 
+    optionLabel="cname" optionGroupLabel="name" optionGroupChildren={['states', 'cities']}
+    className="w-full md:w-14rem" breakpoint="767px" placeholder="Select a City" style={{ minWidth: '14rem' }}  />
+        `,
+        javascript: `
+import React, { useState } from "react";
+import { CascadeSelect } from 'primereact/cascadeselect';
+
+export default function FilledDemo() {
+    const [selectedCity, setSelectedCity] = useState(null);
+    const countries = [
+        {
+            name: 'Australia',
+            code: 'AU',
+            states: [
+                {
+                    name: 'New South Wales',
+                    cities: [
+                        {cname: 'Sydney', code: 'A-SY'},
+                        {cname: 'Newcastle', code: 'A-NE'},
+                        {cname: 'Wollongong', code: 'A-WO'}
+                    ]
+                },
+                {
+                    name: 'Queensland',
+                    cities: [
+                        {cname: 'Brisbane', code: 'A-BR'},
+                        {cname: 'Townsville', code: 'A-TO'}
+                    ]
+                },
+
+            ]
+        },
+        {
+            name: 'Canada',
+            code: 'CA',
+            states: [
+                {
+                    name: 'Quebec',
+                    cities: [
+                        {cname: 'Montreal', code: 'C-MO'},
+                        {cname: 'Quebec City', code: 'C-QU'}
+                    ]
+                },
+                {
+                    name: 'Ontario',
+                    cities: [
+                        {cname: 'Ottawa', code: 'C-OT'},
+                        {cname: 'Toronto', code: 'C-TO'}
+                    ]
+                },
+
+            ]
+        },
+        {
+            name: 'United States',
+            code: 'US',
+            states: [
+                {
+                    name: 'California',
+                    cities: [
+                        {cname: 'Los Angeles', code: 'US-LA'},
+                        {cname: 'San Diego', code: 'US-SD'},
+                        {cname: 'San Francisco', code: 'US-SF'}
+                    ]
+                },
+                {
+                    name: 'Florida',
+                    cities: [
+                        {cname: 'Jacksonville', code: 'US-JA'},
+                        {cname: 'Miami', code: 'US-MI'},
+                        {cname: 'Tampa', code: 'US-TA'},
+                        {cname: 'Orlando', code: 'US-OR'}
+                    ]
+                },
+                {
+                    name: 'Texas',
+                    cities: [
+                        {cname: 'Austin', code: 'US-AU'},
+                        {cname: 'Dallas', code: 'US-DA'},
+                        {cname: 'Houston', code: 'US-HO'}
+                    ]
+                }
+            ]
+        }
+    ];
+
+    return (
+        <div className="card flex justify-content-center">
+            <CascadeSelect variant="filled" value={selectedCity} onChange={(e) => setSelectedCity(e.value)} options={countries} 
+                optionLabel="cname" optionGroupLabel="name" optionGroupChildren={['states', 'cities']}
+                className="w-full md:w-14rem" breakpoint="767px" placeholder="Select a City" style={{ minWidth: '14rem' }}  />
+        </div>
+    )
+}
+        `,
+        typescript: `
+import React, { useState } from "react";
+import { CascadeSelect, CascadeSelectChangeEvent } from 'primereact/cascadeselect';
+
+interface City {
+    cname: string;
+    code: string;
+}
+
+interface CountryState {
+    name: string;
+    cities: City[];
+}
+
+interface Country {
+    name: string;
+    code: string;
+    states: CountryState[];
+}
+
+export default function FilledDemo() {
+    const [selectedCity, setSelectedCity] = useState<City | null>(null);
+    const countries: Country[] = [
+        {
+            name: 'Australia',
+            code: 'AU',
+            states: [
+                {
+                    name: 'New South Wales',
+                    cities: [
+                        {cname: 'Sydney', code: 'A-SY'},
+                        {cname: 'Newcastle', code: 'A-NE'},
+                        {cname: 'Wollongong', code: 'A-WO'}
+                    ]
+                },
+                {
+                    name: 'Queensland',
+                    cities: [
+                        {cname: 'Brisbane', code: 'A-BR'},
+                        {cname: 'Townsville', code: 'A-TO'}
+                    ]
+                },
+
+            ]
+        },
+        {
+            name: 'Canada',
+            code: 'CA',
+            states: [
+                {
+                    name: 'Quebec',
+                    cities: [
+                        {cname: 'Montreal', code: 'C-MO'},
+                        {cname: 'Quebec City', code: 'C-QU'}
+                    ]
+                },
+                {
+                    name: 'Ontario',
+                    cities: [
+                        {cname: 'Ottawa', code: 'C-OT'},
+                        {cname: 'Toronto', code: 'C-TO'}
+                    ]
+                },
+
+            ]
+        },
+        {
+            name: 'United States',
+            code: 'US',
+            states: [
+                {
+                    name: 'California',
+                    cities: [
+                        {cname: 'Los Angeles', code: 'US-LA'},
+                        {cname: 'San Diego', code: 'US-SD'},
+                        {cname: 'San Francisco', code: 'US-SF'}
+                    ]
+                },
+                {
+                    name: 'Florida',
+                    cities: [
+                        {cname: 'Jacksonville', code: 'US-JA'},
+                        {cname: 'Miami', code: 'US-MI'},
+                        {cname: 'Tampa', code: 'US-TA'},
+                        {cname: 'Orlando', code: 'US-OR'}
+                    ]
+                },
+                {
+                    name: 'Texas',
+                    cities: [
+                        {cname: 'Austin', code: 'US-AU'},
+                        {cname: 'Dallas', code: 'US-DA'},
+                        {cname: 'Houston', code: 'US-HO'}
+                    ]
+                }
+            ]
+        }
+    ];
+
+    return (
+        <div className="card flex justify-content-center">
+            <CascadeSelect variant="filled" value={selectedCity} onChange={(e: CascadeSelectChangeEvent) => setSelectedCity(e.value)} options={countries} 
+                optionLabel="cname" optionGroupLabel="name" optionGroupChildren={['states', 'cities']}
+                className="w-full md:w-14rem" breakpoint="767px" placeholder="Select a City" style={{ minWidth: '14rem' }}  />
+        </div>
+    )
+}
+        `
+    };
+
+    return (
+        <>
+            <DocSectionText {...props}>
+                <p>
+                    Specify the <i>variant</i> property as <i>filled</i> to display the component with a higher visual emphasis than the default <i>outlined</i> style.
+                </p>
+            </DocSectionText>
+            <div className="card flex justify-content-center">
+                <CascadeSelect
+                    variant="filled"
+                    value={selectedCity}
+                    onChange={(e) => setSelectedCity(e.value)}
+                    options={countries}
+                    optionLabel="cname"
+                    optionGroupLabel="name"
+                    optionGroupChildren={['states', 'cities']}
+                    className="w-full md:w-14rem"
+                    breakpoint="767px"
+                    placeholder="Select a City"
+                    style={{ minWidth: '14rem' }}
+                />
+            </div>
+            <DocSectionCode code={code} />
+        </>
+    );
+}
