--- components/doc/cascadeselect/form/formikdoc.js
@@ -1,470 +1,0 @@
-import { DocSectionCode } from '@/components/doc/common/docsectioncode';
-import { DocSectionText } from '@/components/doc/common/docsectiontext';
-import { Button } from '@/components/lib/button/Button';
-import { CascadeSelect } from '@/components/lib/cascadeselect/CascadeSelect';
-import { Toast } from '@/components/lib/toast/Toast';
-import { classNames } from '@/components/lib/utils/Utils';
-import { useFormik } from 'formik';
-import { useRef } from 'react';
-
-export function FormikDoc(props) {
-    const toast = useRef(null);
-    const countries = [
-        {
-            name: 'Australia',
-            code: 'AU',
-            states: [
-                {
-                    name: 'New South Wales',
-                    cities: [
-                        { cname: 'Sydney', code: 'A-SY' },
-                        { cname: 'Newcastle', code: 'A-NE' },
-                        { cname: 'Wollongong', code: 'A-WO' }
-                    ]
-                },
-                {
-                    name: 'Queensland',
-                    cities: [
-                        { cname: 'Brisbane', code: 'A-BR' },
-                        { cname: 'Townsville', code: 'A-TO' }
-                    ]
-                }
-            ]
-        },
-        {
-            name: 'Canada',
-            code: 'CA',
-            states: [
-                {
-                    name: 'Quebec',
-                    cities: [
-                        { cname: 'Montreal', code: 'C-MO' },
-                        { cname: 'Quebec City', code: 'C-QU' }
-                    ]
-                },
-                {
-                    name: 'Ontario',
-                    cities: [
-                        { cname: 'Ottawa', code: 'C-OT' },
-                        { cname: 'Toronto', code: 'C-TO' }
-                    ]
-                }
-            ]
-        },
-        {
-            name: 'United States',
-            code: 'US',
-            states: [
-                {
-                    name: 'California',
-                    cities: [
-                        { cname: 'Los Angeles', code: 'US-LA' },
-                        { cname: 'San Diego', code: 'US-SD' },
-                        { cname: 'San Francisco', code: 'US-SF' }
-                    ]
-                },
-                {
-                    name: 'Florida',
-                    cities: [
-                        { cname: 'Jacksonville', code: 'US-JA' },
-                        { cname: 'Miami', code: 'US-MI' },
-                        { cname: 'Tampa', code: 'US-TA' },
-                        { cname: 'Orlando', code: 'US-OR' }
-                    ]
-                },
-                {
-                    name: 'Texas',
-                    cities: [
-                        { cname: 'Austin', code: 'US-AU' },
-                        { cname: 'Dallas', code: 'US-DA' },
-                        { cname: 'Houston', code: 'US-HO' }
-                    ]
-                }
-            ]
-        }
-    ];
-
-    const show = (data) => {
-        toast.current.show({ severity: 'success', summary: 'Form Submitted', detail: `${data.city.cname}` });
-    };
-
-    const formik = useFormik({
-        initialValues: {
-            city: null
-        },
-        validate: (data) => {
-            let errors = {};
-
-            if (!data.city) {
-                errors.city = 'City is required.';
-            }
-
-            return errors;
-        },
-        onSubmit: (data) => {
-            data && show(data);
-            formik.resetForm();
-        }
-    });
-
-    const isFormFieldInvalid = (name) => !!(formik.touched[name] && formik.errors[name]);
-
-    const getFormErrorMessage = (name) => {
-        return isFormFieldInvalid(name) ? <small className="p-error">{formik.errors[name]}</small> : <small className="p-error">&nbsp;</small>;
-    };
-
-    const code = {
-        basic: `
-<Toast ref={toast} />
-<CascadeSelect
-    inputId="cc_city"
-    name="city"
-    value={formik.values.city}
-    options={countries}
-    optionLabel="cname"
-    optionGroupLabel="name"
-    optionGroupChildren={['states', 'cities']}
-    style={{ minWidth: '14rem' }}
-    breakpoint="767px"
-    placeholder="Select a City"
-    onChange={(e) => {
-        formik.setFieldValue('city', e.value);
-    }}
-/>
-{getFormErrorMessage('city')}
-<Button type="submit" label="Submit" />
-        `,
-        javascript: `
-import React, { useRef } from "react";
-import { useFormik } from 'formik';
-import { CascadeSelect } from 'primereact/cascadeselect';
-import { Button } from 'primereact/button';
-import { Toast } from 'primereact/toast';
-
-export default function FormikDoc() {
-    const toast = useRef(null);
-    const countries = [
-        {
-            name: 'Australia',
-            code: 'AU',
-            states: [
-                {
-                    name: 'New South Wales',
-                    cities: [
-                        { cname: 'Sydney', code: 'A-SY' },
-                        { cname: 'Newcastle', code: 'A-NE' },
-                        { cname: 'Wollongong', code: 'A-WO' }
-                    ]
-                },
-                {
-                    name: 'Queensland',
-                    cities: [
-                        { cname: 'Brisbane', code: 'A-BR' },
-                        { cname: 'Townsville', code: 'A-TO' }
-                    ]
-                }
-            ]
-        },
-        {
-            name: 'Canada',
-            code: 'CA',
-            states: [
-                {
-                    name: 'Quebec',
-                    cities: [
-                        { cname: 'Montreal', code: 'C-MO' },
-                        { cname: 'Quebec City', code: 'C-QU' }
-                    ]
-                },
-                {
-                    name: 'Ontario',
-                    cities: [
-                        { cname: 'Ottawa', code: 'C-OT' },
-                        { cname: 'Toronto', code: 'C-TO' }
-                    ]
-                }
-            ]
-        },
-        {
-            name: 'United States',
-            code: 'US',
-            states: [
-                {
-                    name: 'California',
-                    cities: [
-                        { cname: 'Los Angeles', code: 'US-LA' },
-                        { cname: 'San Diego', code: 'US-SD' },
-                        { cname: 'San Francisco', code: 'US-SF' }
-                    ]
-                },
-                {
-                    name: 'Florida',
-                    cities: [
-                        { cname: 'Jacksonville', code: 'US-JA' },
-                        { cname: 'Miami', code: 'US-MI' },
-                        { cname: 'Tampa', code: 'US-TA' },
-                        { cname: 'Orlando', code: 'US-OR' }
-                    ]
-                },
-                {
-                    name: 'Texas',
-                    cities: [
-                        { cname: 'Austin', code: 'US-AU' },
-                        { cname: 'Dallas', code: 'US-DA' },
-                        { cname: 'Houston', code: 'US-HO' }
-                    ]
-                }
-            ]
-        }
-    ];
-
-    const show = (data) => {
-        toast.current.show({ severity: 'success', summary: 'Form Submitted', detail: \`\${data.city.cname}\`});
-    };
-
-    const formik = useFormik({
-        initialValues: {
-            city: null
-        },
-        validate: (data) => {
-            let errors = {};
-
-            if (!data.city) {
-                errors.city = 'City is required.';
-            }
-
-            return errors;
-        },
-        onSubmit: (data) => {
-            data && show(data);
-            formik.resetForm();
-        }
-    });
-
-    const isFormFieldInvalid = (name) => !!(formik.touched[name] && formik.errors[name]);
-
-    const getFormErrorMessage = (name) => {
-        return isFormFieldInvalid(name) ? <small className="p-error">{formik.errors[name]}</small> : <small className="p-error">&nbsp;</small>;
-
-    };
-
-    return (
-        <div className="card flex justify-content-center">
-            <form onSubmit={formik.handleSubmit} className="flex flex-column gap-2">
-                <label htmlFor="cc_city">Select a city</label>
-                <Toast ref={toast} />
-                <CascadeSelect
-                    inputId="cc_city"
-                    name="city"
-                    value={formik.values.city}
-                    options={countries}
-                    optionLabel="cname"
-                    optionGroupLabel="name"
-                    optionGroupChildren={['states', 'cities']}
-                    style={{ minWidth: '14rem' }}
-                    breakpoint="767px"
-                    placeholder="Select a City"
-                    onChange={(e) => {
-                        formik.setFieldValue('city', e.value);
-                    }}
-                />
-                {getFormErrorMessage('city')}
-                <Button type="submit" label="Submit" />
-            </form>
-        </div>
-    )
-}
-        `,
-        typescript: `
-import React, { useRef } from "react";
-import { useFormik } from 'formik';
-import { CascadeSelect, CascadeSelectChangeEvent } from 'primereact/cascadeselect';
-import { Button } from 'primereact/button';
-import { Toast } from 'primereact/toast';
-
-interface City {
-    cname: string;
-    code: string;
-}
-
-interface CountryState {
-    name: string;
-    cities: City[];
-}
-
-interface Country {
-    name: string;
-    code: string;
-    states: CountryState[];
-}
-
-export default function FormikDoc() {
-    const toast = useRef(null);
-    const countries: Country[] = [
-        {
-            name: 'Australia',
-            code: 'AU',
-            states: [
-                {
-                    name: 'New South Wales',
-                    cities: [
-                        { cname: 'Sydney', code: 'A-SY' },
-                        { cname: 'Newcastle', code: 'A-NE' },
-                        { cname: 'Wollongong', code: 'A-WO' }
-                    ]
-                },
-                {
-                    name: 'Queensland',
-                    cities: [
-                        { cname: 'Brisbane', code: 'A-BR' },
-                        { cname: 'Townsville', code: 'A-TO' }
-                    ]
-                }
-            ]
-        },
-        {
-            name: 'Canada',
-            code: 'CA',
-            states: [
-                {
-                    name: 'Quebec',
-                    cities: [
-                        { cname: 'Montreal', code: 'C-MO' },
-                        { cname: 'Quebec City', code: 'C-QU' }
-                    ]
-                },
-                {
-                    name: 'Ontario',
-                    cities: [
-                        { cname: 'Ottawa', code: 'C-OT' },
-                        { cname: 'Toronto', code: 'C-TO' }
-                    ]
-                }
-            ]
-        },
-        {
-            name: 'United States',
-            code: 'US',
-            states: [
-                {
-                    name: 'California',
-                    cities: [
-                        { cname: 'Los Angeles', code: 'US-LA' },
-                        { cname: 'San Diego', code: 'US-SD' },
-                        { cname: 'San Francisco', code: 'US-SF' }
-                    ]
-                },
-                {
-                    name: 'Florida',
-                    cities: [
-                        { cname: 'Jacksonville', code: 'US-JA' },
-                        { cname: 'Miami', code: 'US-MI' },
-                        { cname: 'Tampa', code: 'US-TA' },
-                        { cname: 'Orlando', code: 'US-OR' }
-                    ]
-                },
-                {
-                    name: 'Texas',
-                    cities: [
-                        { cname: 'Austin', code: 'US-AU' },
-                        { cname: 'Dallas', code: 'US-DA' },
-                        { cname: 'Houston', code: 'US-HO' }
-                    ]
-                }
-            ]
-        }
-    ];
-
-    const show = (data) => {
-        toast.current.show({ severity: 'success', summary: 'Form Submitted', detail: \`\${data.city.cname}\`});
-    };
-
-    const formik = useFormik({
-        initialValues: {
-            city: null
-        },
-        validate: (data) => {
-            let errors = {};
-
-            if (!data.city) {
-                errors.city = 'City is required.';
-            }
-
-            return errors;
-        },
-        onSubmit: (data) => {
-            data && show(data);
-            formik.resetForm();
-        }
-    });
-
-    const isFormFieldInvalid = (name) => !!(formik.touched[name] && formik.errors[name]);
-
-    const getFormErrorMessage = (name) => {
-        return isFormFieldInvalid(name) ? <small className="p-error">{formik.errors[name]}</small> : <small className="p-error">&nbsp;</small>;
-
-    };
-
-    return (
-        <div className="card flex justify-content-center">
-            <form onSubmit={formik.handleSubmit} className="flex flex-column gap-2">
-                <label htmlFor="cc_city">Select a city</label>
-                <Toast ref={toast} />
-                <CascadeSelect
-                    inputId="cc_city"
-                    name="city"
-                    value={formik.values.city}
-                    options={countries}
-                    optionLabel="cname"
-                    optionGroupLabel="name"
-                    optionGroupChildren={['states', 'cities']}
-                    style={{ minWidth: '14rem' }}
-                    breakpoint="767px"
-                    placeholder="Select a City"
-                    onChange={(e: CascadeSelectChangeEvent) => {
-                        formik.setFieldValue('city', e.value);
-                    }}
-                />
-                {getFormErrorMessage('city')}
-                <Button type="submit" label="Submit" />
-            </form>
-        </div>
-    )
-}
-        `
-    };
-
-    return (
-        <>
-            <DocSectionText {...props}>
-                <p>
-                    <a href="https://formik.org/">Formik</a> is a popular library for handling forms in React.
-                </p>
-            </DocSectionText>
-            <div className="card flex justify-content-center">
-                <form onSubmit={formik.handleSubmit} className="flex flex-column gap-2">
-                    <Toast ref={toast} />
-                    <CascadeSelect
-                        inputId="cc_city"
-                        name="city"
-                        value={formik.values.city}
-                        options={countries}
-                        optionLabel="cname"
-                        optionGroupLabel="name"
-                        optionGroupChildren={['states', 'cities']}
-                        style={{ minWidth: '14rem' }}
-                        breakpoint="767px"
-                        placeholder="Select a City"
-                        className={classNames({ 'p-invalid': isFormFieldInvalid('city') })}
-                        onChange={(e) => {
-                            formik.setFieldValue('city', e.value);
-                        }}
-                    />
-                    {getFormErrorMessage('city')}
-                    <Button type="submit" label="Submit" />
-                </form>
-            </div>
-            <DocSectionCode code={code} dependencies={{ formik: '^2.2.6' }} />
-        </>
-    );
-}
