--- components/doc/cascadeselect/form/hookformdoc.js
@@ -1,478 +1,0 @@
-import { DocSectionCode } from '@/components/doc/common/docsectioncode';
-import { DocSectionText } from '@/components/doc/common/docsectiontext';
-import { Button } from '@/components/lib/button/Button';
-import { CascadeSelect } from '@/components/lib/cascadeselect/CascadeSelect';
-import { Toast } from '@/components/lib/toast/Toast';
-import { classNames } from '@/components/lib/utils/Utils';
-import { useRef } from 'react';
-import { Controller, useForm } from 'react-hook-form';
-
-export function HookFormDoc(props) {
-    const toast = useRef(null);
-    const countries = [
-        {
-            name: 'Australia',
-            code: 'AU',
-            states: [
-                {
-                    name: 'New South Wales',
-                    cities: [
-                        { cname: 'Sydney', code: 'A-SY' },
-                        { cname: 'Newcastle', code: 'A-NE' },
-                        { cname: 'Wollongong', code: 'A-WO' }
-                    ]
-                },
-                {
-                    name: 'Queensland',
-                    cities: [
-                        { cname: 'Brisbane', code: 'A-BR' },
-                        { cname: 'Townsville', code: 'A-TO' }
-                    ]
-                }
-            ]
-        },
-        {
-            name: 'Canada',
-            code: 'CA',
-            states: [
-                {
-                    name: 'Quebec',
-                    cities: [
-                        { cname: 'Montreal', code: 'C-MO' },
-                        { cname: 'Quebec City', code: 'C-QU' }
-                    ]
-                },
-                {
-                    name: 'Ontario',
-                    cities: [
-                        { cname: 'Ottawa', code: 'C-OT' },
-                        { cname: 'Toronto', code: 'C-TO' }
-                    ]
-                }
-            ]
-        },
-        {
-            name: 'United States',
-            code: 'US',
-            states: [
-                {
-                    name: 'California',
-                    cities: [
-                        { cname: 'Los Angeles', code: 'US-LA' },
-                        { cname: 'San Diego', code: 'US-SD' },
-                        { cname: 'San Francisco', code: 'US-SF' }
-                    ]
-                },
-                {
-                    name: 'Florida',
-                    cities: [
-                        { cname: 'Jacksonville', code: 'US-JA' },
-                        { cname: 'Miami', code: 'US-MI' },
-                        { cname: 'Tampa', code: 'US-TA' },
-                        { cname: 'Orlando', code: 'US-OR' }
-                    ]
-                },
-                {
-                    name: 'Texas',
-                    cities: [
-                        { cname: 'Austin', code: 'US-AU' },
-                        { cname: 'Dallas', code: 'US-DA' },
-                        { cname: 'Houston', code: 'US-HO' }
-                    ]
-                }
-            ]
-        }
-    ];
-
-    const show = (data) => {
-        toast.current.show({ severity: 'success', summary: 'Form Submitted', detail: `${data.city.cname}` });
-    };
-
-    const defaultValues = {
-        city: null
-    };
-
-    const {
-        control,
-        formState: { errors },
-        handleSubmit,
-        reset
-    } = useForm({ defaultValues });
-
-    const onSubmit = (data) => {
-        data.city && show(data);
-
-        reset();
-    };
-
-    const getFormErrorMessage = (name) => {
-        return errors[name] ? <small className="p-error">{errors[name].message}</small> : <small className="p-error">&nbsp;</small>;
-    };
-
-    const code = {
-        basic: `
-<Toast ref={toast} />
-<Controller
-    name="city"
-    control={control}
-    rules={{ required: 'City is required.' }}
-    render={({ field }) => (
-        <CascadeSelect
-            id={field.name}
-            name="city"
-            value={field.value}
-            options={countries}
-            optionLabel="cname"
-            optionGroupLabel="name"
-            optionGroupChildren={['states', 'cities']}
-            style={{ minWidth: '14rem' }}
-            breakpoint="767px"
-            placeholder="Select a City"
-            onChange={(e) => field.onChange(e.value)}
-        />
-    )}
-/>
-{getFormErrorMessage('city')}
-<Button type="submit" label="Submit" />
-        `,
-        javascript: `
-import React, { useRef } from "react";
-import { useForm, Controller } from 'react-hook-form';
-import { CascadeSelect } from 'primereact/cascadeselect';
-import { Button } from 'primereact/button';
-import { Toast } from 'primereact/toast';
-import { classNames } from 'primereact/utils';
-
-export default function HookFormDoc() {
-    const toast = useRef(null);
-    const countries = [
-        {
-            name: 'Australia',
-            code: 'AU',
-            states: [
-                {
-                    name: 'New South Wales',
-                    cities: [
-                        { cname: 'Sydney', code: 'A-SY' },
-                        { cname: 'Newcastle', code: 'A-NE' },
-                        { cname: 'Wollongong', code: 'A-WO' }
-                    ]
-                },
-                {
-                    name: 'Queensland',
-                    cities: [
-                        { cname: 'Brisbane', code: 'A-BR' },
-                        { cname: 'Townsville', code: 'A-TO' }
-                    ]
-                }
-            ]
-        },
-        {
-            name: 'Canada',
-            code: 'CA',
-            states: [
-                {
-                    name: 'Quebec',
-                    cities: [
-                        { cname: 'Montreal', code: 'C-MO' },
-                        { cname: 'Quebec City', code: 'C-QU' }
-                    ]
-                },
-                {
-                    name: 'Ontario',
-                    cities: [
-                        { cname: 'Ottawa', code: 'C-OT' },
-                        { cname: 'Toronto', code: 'C-TO' }
-                    ]
-                }
-            ]
-        },
-        {
-            name: 'United States',
-            code: 'US',
-            states: [
-                {
-                    name: 'California',
-                    cities: [
-                        { cname: 'Los Angeles', code: 'US-LA' },
-                        { cname: 'San Diego', code: 'US-SD' },
-                        { cname: 'San Francisco', code: 'US-SF' }
-                    ]
-                },
-                {
-                    name: 'Florida',
-                    cities: [
-                        { cname: 'Jacksonville', code: 'US-JA' },
-                        { cname: 'Miami', code: 'US-MI' },
-                        { cname: 'Tampa', code: 'US-TA' },
-                        { cname: 'Orlando', code: 'US-OR' }
-                    ]
-                },
-                {
-                    name: 'Texas',
-                    cities: [
-                        { cname: 'Austin', code: 'US-AU' },
-                        { cname: 'Dallas', code: 'US-DA' },
-                        { cname: 'Houston', code: 'US-HO' }
-                    ]
-                }
-            ]
-        }
-    ];
-
-    const show = (data) => {
-        toast.current.show({ severity: 'success', summary: 'Form Submitted', detail: \`\${data.city.cname}\` });
-    };
-
-    const defaultValues = {
-        city: null
-    };
-
-    const {
-        control,
-        formState: { errors },
-        handleSubmit,
-        reset
-    } = useForm({ defaultValues });
-
-    const onSubmit = (data) => {
-        data.city && show(data);
-
-        reset();
-    };
-
-    const getFormErrorMessage = (name) => {
-        return errors[name] ? <small className="p-error">{errors[name].message}</small> : <small className="p-error">&nbsp;</small>;
-    };
-
-    return (
-        <div className="card flex justify-content-center">
-            <form onSubmit={handleSubmit(onSubmit)} className="flex flex-column gap-2">
-                <Toast ref={toast} />
-                <Controller
-                    name="city"
-                    control={control}
-                    rules={{ required: 'City is required.' }}
-                    render={({ field, fieldState }) => (
-                        <CascadeSelect
-                            id={field.name}
-                            name="city"
-                            value={field.value}
-                            options={countries}
-                            optionLabel="cname"
-                            optionGroupLabel="name"
-                            optionGroupChildren={['states', 'cities']}
-                            style={{ minWidth: '14rem' }}
-                            breakpoint="767px"
-                            placeholder="Select a City"
-                            className={classNames({ 'p-invalid': fieldState.error })}
-                            onChange={(e) => field.onChange(e.value)}
-                        />
-                    )}
-                />
-                {getFormErrorMessage('city')}
-                <Button type="submit" label="Submit" />
-            </form>
-        </div>
-    )
-}
-        `,
-        typescript: `
-import React, { useRef } from "react";
-import { useForm, Controller } from 'react-hook-form';
-import { CascadeSelect, CascadeSelectChangeEvent } from 'primereact/cascadeselect';
-import { Button } from 'primereact/button';
-import { Toast } from 'primereact/toast';
-import { classNames } from 'primereact/utils';
-
-interface City {
-    cname: string;
-    code: string;
-}
-
-interface CountryState {
-    name: string;
-    cities: City[];
-}
-
-interface Country {
-    name: string;
-    code: string;
-    states: CountryState[];
-}
-
-export default function HookFormDoc() {
-    const toast = useRef(null);
-    const countries: Country[] = [
-        {
-            name: 'Australia',
-            code: 'AU',
-            states: [
-                {
-                    name: 'New South Wales',
-                    cities: [
-                        { cname: 'Sydney', code: 'A-SY' },
-                        { cname: 'Newcastle', code: 'A-NE' },
-                        { cname: 'Wollongong', code: 'A-WO' }
-                    ]
-                },
-                {
-                    name: 'Queensland',
-                    cities: [
-                        { cname: 'Brisbane', code: 'A-BR' },
-                        { cname: 'Townsville', code: 'A-TO' }
-                    ]
-                }
-            ]
-        },
-        {
-            name: 'Canada',
-            code: 'CA',
-            states: [
-                {
-                    name: 'Quebec',
-                    cities: [
-                        { cname: 'Montreal', code: 'C-MO' },
-                        { cname: 'Quebec City', code: 'C-QU' }
-                    ]
-                },
-                {
-                    name: 'Ontario',
-                    cities: [
-                        { cname: 'Ottawa', code: 'C-OT' },
-                        { cname: 'Toronto', code: 'C-TO' }
-                    ]
-                }
-            ]
-        },
-        {
-            name: 'United States',
-            code: 'US',
-            states: [
-                {
-                    name: 'California',
-                    cities: [
-                        { cname: 'Los Angeles', code: 'US-LA' },
-                        { cname: 'San Diego', code: 'US-SD' },
-                        { cname: 'San Francisco', code: 'US-SF' }
-                    ]
-                },
-                {
-                    name: 'Florida',
-                    cities: [
-                        { cname: 'Jacksonville', code: 'US-JA' },
-                        { cname: 'Miami', code: 'US-MI' },
-                        { cname: 'Tampa', code: 'US-TA' },
-                        { cname: 'Orlando', code: 'US-OR' }
-                    ]
-                },
-                {
-                    name: 'Texas',
-                    cities: [
-                        { cname: 'Austin', code: 'US-AU' },
-                        { cname: 'Dallas', code: 'US-DA' },
-                        { cname: 'Houston', code: 'US-HO' }
-                    ]
-                }
-            ]
-        }
-    ];
-
-    const show = (data) => {
-        toast.current.show({ severity: 'success', summary: 'Form Submitted', detail: \`\${data.city.cname}\` });
-    };
-
-    const defaultValues = {
-        city: null
-    };
-
-    const {
-        control,
-        formState: { errors },
-        handleSubmit,
-        reset
-    } = useForm({ defaultValues });
-
-    const onSubmit = (data) => {
-        data.city && show(data);
-
-        reset();
-    };
-
-    const getFormErrorMessage = (name) => {
-        return errors[name] ? <small className="p-error">{errors[name].message}</small> : <small className="p-error">&nbsp;</small>;
-    };
-
-    return (
-        <div className="card flex justify-content-center">
-            <form onSubmit={handleSubmit(onSubmit)} className="flex flex-column gap-2">
-                <Toast ref={toast} />
-                <Controller
-                    name="city"
-                    control={control}
-                    rules={{ required: 'City is required.' }}
-                    render={({ field, fieldState }) => (
-                        <CascadeSelect
-                            id={field.name}
-                            name="city"
-                            value={field.value}
-                            options={countries}
-                            optionLabel="cname"
-                            optionGroupLabel="name"
-                            optionGroupChildren={['states', 'cities']}
-                            style={{ minWidth: '14rem' }}
-                            breakpoint="767px"
-                            placeholder="Select a City"
-                            className={classNames({ 'p-invalid': fieldState.error })}
-                            onChange={(e: CascadeSelectChangeEvent) => field.onChange(e.value)}
-                        />
-                    )}
-                />
-                {getFormErrorMessage('city')}
-                <Button type="submit" label="Submit" />
-            </form>
-        </div>
-    )
-}
-        `
-    };
-
-    return (
-        <>
-            <DocSectionText {...props}>
-                <p>
-                    <a href="https://react-hook-form.com/">React Hook Form</a> is another popular React library to handle forms.
-                </p>
-            </DocSectionText>
-            <div className="card flex justify-content-center">
-                <form onSubmit={handleSubmit(onSubmit)} className="flex flex-column gap-2">
-                    <Toast ref={toast} />
-                    <Controller
-                        name="city"
-                        control={control}
-                        rules={{ required: 'City is required.' }}
-                        render={({ field, fieldState }) => (
-                            <CascadeSelect
-                                id={field.name}
-                                name="city"
-                                value={field.value}
-                                options={countries}
-                                optionLabel={'cname'}
-                                optionGroupLabel={'name'}
-                                optionGroupChildren={['states', 'cities']}
-                                style={{ minWidth: '14rem' }}
-                                breakpoint="767px"
-                                placeholder={'Select a City'}
-                                className={classNames({ 'p-invalid': fieldState.error })}
-                                onChange={(e) => field.onChange(e.value)}
-                            />
-                        )}
-                    />
-                    {getFormErrorMessage('city')}
-                    <Button type="submit" label="Submit" />
-                </form>
-            </div>
-            <DocSectionCode code={code} dependencies={{ 'react-hook-form': '^7.39.4' }} />
-        </>
-    );
-}
